feat: update settings implementation
BREAKING CHANGE: Settings implementation has been updated

The path for settings has been updated from 'values/xml/settings.xml' to 'kotlin/.../settings/...'. This change may affect existing configurations and tools that rely on the previous path.

To update your configurations and tools, please make the necessary changes to reflect the new path. I'm apologize for any inconvenience this may cause.

+++ app/src/main/kotlin/org/cosmicide/rewrite/treeview/ViewBinder.kt
@@ -67,7 +67,7 @@ class ViewBinder(
     ) {
         with(holder.itemView.findViewById<Space>(R.id.space)) {
             updateLayoutParams<ViewGroup.MarginLayoutParams> {
-                width = node.depth * context.getDip(22f)
+                width = node.depth * context.getDip(22f).toInt()
             }
         }
